Deptreevis file (#10)
This PR introduces the `DepTreeVisFile` Struct. This file acts as a
representation for all relevant files read by `deptreevis`.

I don't think there is a point in distinguishing files based on their
extension (`.py`, `.rs`, etc) because a file is a file, how it is
handled is determined by the respective walker.

No tests for this - not enough functionality yet :)

---------

Co-authored-by: sudoPom <sudopom@github.com>